Added support for building common with -Wall. ( only 5-6 errors left now :) )

+++ src/common/doc_cache.c
@@ -7,8 +7,8 @@
 unsigned int sign_cache_params(docid doc_id, char *subname, time_t time) {
 	char data[maxSubnameLength + 128];
 
-	snprintf(data, sizeof data, "%u %s %u %s", 
-		doc_id, subname, time, DOC_CACHE_SALT);
+	snprintf(data, sizeof data, "%u %s %lld %s", 
+		doc_id, subname, (long long)time, DOC_CACHE_SALT);
 
 	return crc32boitho(data);
 }
